refactored app.js
