+++ page/src/App.css
@@ -12,7 +12,7 @@ h1 {
 }
 
 .input-section {
-  width: 50%;
+  width: 80;
   margin-bottom: 20px;
 }
 
@@ -41,7 +41,7 @@ h1 {
 }
 
 .journal-display {
-  width: 50%;
+  width: 80%;
   padding: 20px;
   border-radius: 8px;
   border: 1px solid #ccc;
